requirements.txt
pandas==2.2.2
requests==2.32.3
pyyaml==6.0.2
openpyxl==3.1.5
xlrd==2.0.1
